Fix European station order (public broadcasters first, then commercial)
In the absence of a static frequency, the order is generally public broadcasters first (or if that does not exist, the station that takes that role (e.g. RTL in Luxembourg), and then the commercial stations in alphabetical order. The public stations are in order of how they are generally represented, so not necessarily alphabetical. E.g., Vikerraadio comes before Raadio 2